requirements.txt
faiss-cpu 
sentence-transformers 
fastapi 
uvicorn 
markdown-it-py
streamlit
torch